refactor(client): merge files for SubContent input
